Basic features are now complete.
Remaining to do is to document the code, and possibly clean it up in
places (when I've had more than a couple hours of sleep).

It's now going to serve as a tutorial for basic platformer + loading a
FlxTilemap from a color bitmap (an upcoming feature in flixel). Perhaps
if it is useful to users, I can add more features upon request.